Add database handling for repositories and manifests (#2)
This commit includes the creation of repositories and manifests tables
in the database schema in `schema.sql` file. Furthermore, it modifies
the `registry.js` file to handle data insertions, updates, and deletions
in these tables. The `index.js` file has been updated to import the
database, and configuration has been added to `wrangler.toml` to specify
the database binding and ID.